build: fix linkage issues related to pixi/conda environments and cudarc (#13)
The `cupti` feature I introduced in `cudarc` depends on the `driver` and
`runtime` features, because CUPTI has a lot of APIs that expose CUDA
driver and runtime types, for which the definitions of the `driver` and
`runtime` modules are reused.

(An alternative seen in some other parts of the crate are to let bindgen
generate duplicate definitions, but I didn't want that, because it is
both confusing and it would encumber a workflow where you would do
something with the driver/runtime based on some CUPTI interaction.)

When enabling the `driver` and `runtime` features of `cudarc`, their
`result` (and `safe`) submodules wrap around these libraries, so they
have to link to them.

In order to do so, `cudarc` adds typical CUDA toolkit install paths
("standard locations") to the linker library search paths as a part of
its build script.

However, this is problematic when working in a virtual env with the CUDA
toolkit from e.g. Conda or Pixi, which doesn't come with `libcuda.so`.
Typically, the driver is dynamically loaded by e.g. `cudart` or `cupti`.
This is an option that cudarc does provide.

Some issues, however, are:
1. cudarc dynamic loading errors out if certain symbols aren't found,
even if your application doesn't use them. this isn't always truly a
problem, e.g. if you have a somewhat older system install of cuda but a
newer cuda toolkit in your env which doesn't use any new driver api
function
2. adding "standard locations" to the library search paths pointing to
system directories will cause all sorts of nasty linker errors because
the linker picks up system installed gcc-related libraries, e.g.:
```
lib/Scrt1.o: in function `_start':
          (.text+0x16): undefined reference to `__libc_csu_fini'
```

I think the best solution is for `cudarc` to:
1. Only attempt to infer the cuda toolkit install path if no
`CUDA_HOME`-like environment variables were set during the build, so
when building in a virtual env you can point to e.g. `$CONDA_PREFIX`.
Don't *always* use "standard locations" to pollute ld/lld library search
paths. `CUDA_HOME` seems to be a common variable among quite a lot of
other projects.
2. Split out the `driver` and `runtime` feature into two, one for just
exposing the `sys` module types so it doesn't require linking, and one
for the wrappers that do require linking.
3. Allow dynamic loading to the driver with missing symbols, erroring
out only when an unfound symbol is actually used, rather than error if
not all symbol are found. This would ensure you can dynamically load an
older driver or other lib as long as you don't use any new
functionality.

To-do, WIP on:
https://github.com/johanpel/cudarc/tree/optional-driver-linkage
- [x] don't use "standard locations" if CUDA_HOME is supplied
- [x] split out exposing `driver` and `runtime` types behind
`driver-sys` and `runtime-sys` flags, the linking required stuff under
`driver` and `runtime`
- [x] don't error the entire program when some (potentially unused)
function pointers can't be found upon dynamic loading, only error out
when such a function is being used.